requirements.txt
phidata 
yfinance
duckduckgo-search
groq
openai
streamlit
pandas
plotly
googlesearch-python
pycountry
polygon-api-client
python-dotenv
requests
datetime
supabase
